examples/old_examples/grasping/requirements.txt
numpy>=1.16.0
PyYAML>=3.11
scipy>=1.2.0
matplotlib>=2.2.2
Pillow>=5.1.0
opencv-python==4.0.0.21
open3d_python==0.5.0.0
bezier==0.9.0
scikit-learn==0.20.3
absl-py==0.7.1
pytest-cov==2.6.1
pytest-html==1.20.0
yacs==0.1.6
guzzle-sphinx-theme==0.7.11
mock==3.0.5
torch==0.3.1
torchvision==0.2.0